refactor: requestWeatherImage메서드의 적용 대상 변경
- UIImageView의 extension으로 되어있던 메서드를 NetworkManager의 메서드로 변경
