+++ Diary/Network/NetworkManager.swift
@@ -5,7 +5,7 @@
 //  Created by unchain, 웡빙 on 2022/08/31.
 //
 
-import Foundation
+import UIKit
 
 final class NetworkManager {
     static let shared = NetworkManager()
@@ -39,6 +39,26 @@ final class NetworkManager {
                 return
             }
             completion(fetchedData)
+        }
+        dataTask.resume()
+    }
+    
+    func requestWeatherImage(iconId: String, iconImage: UIImageView) {
+        guard let url = URL(string: "http://openweathermap.org/img/wn/\(iconId)@2x.png") else { return }
+
+        let dataTask = URLSession.shared.dataTask(with: url) { data, response, error in
+            guard error == nil else { return }
+            let successRange = 200..<300
+            guard let statusCode = (response as? HTTPURLResponse)?.statusCode,
+                  successRange.contains(statusCode) else {
+                return
+            }
+            guard let data = data, let image = UIImage(data: data) else {
+                return
+            }
+            DispatchQueue.main.async {
+                iconImage.image = image
+            }
         }
         dataTask.resume()
     }
